database support feature - step 2 finished variable number deleted from class Task

+++ src/main/java/org/itlabel/cabinet/service/impl/TaskServiceImpl.java
@@ -24,8 +24,9 @@ public class TaskServiceImpl implements TaskService {
     @Override
     @Transactional
     public void deleteTask(Long id) {
-        //TODO - проверить существование Task
-        taskRepository.deleteById(id);
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+        }
     }
 
     @Override
